+++ TxtKudeatzailea.java
@@ -167,91 +167,91 @@ public class TxtKudeatzailea {
     }
 
     private static void txtFitxategiaGehitu() {
-    String fileName, path;
-    Scanner sc = new Gehigarriak().in;
+        String fileName, path;
+        Scanner sc = new Gehigarriak().in;
 
-    System.out.println("TXT fitxategian datuak gehitu");
-    txtFitxategiakBistaratu();
+        System.out.println("TXT fitxategian datuak gehitu");
+        txtFitxategiakBistaratu();
 
-    // Fitxategia aukeratu
-    System.out.print(
-            Gehigarriak.Horia + "Zein fitxategiri datuak gehitu nahi dizkiozu? Sartu izena: " + Gehigarriak.RESET);
-    fileName = sc.next();
-    fileName = Filtroak.removeSpaces(fileName);
-    path = "./fitxategiak/txt/" + fileName + ".txt";
+        // Fitxategia aukeratu
+        System.out.print(
+                Gehigarriak.Horia + "Zein fitxategiri datuak gehitu nahi dizkiozu? Sartu izena: " + Gehigarriak.RESET);
+        fileName = sc.next();
+        fileName = Filtroak.removeSpaces(fileName);
+        path = "./fitxategiak/txt/" + fileName + ".txt";
 
-    sc.nextLine(); // Scanner buffer garbitu
-    String nan, izena, abizena, adinaStr, helbidea;
-    int adina;
+        sc.nextLine(); // Scanner buffer garbitu
+        String nan, izena, abizena, adinaStr, helbidea;
+        int adina;
 
-    // NAN balidazioa + existitzen den ala ez
-    do {
-        System.out.print("NAN (8 zenbaki + 1 letra): ");
-        nan = sc.nextLine();
-        if (!Filtroak.isDNI(nan)) {
+        // NAN balidazioa + existitzen den ala ez
+        do {
+            System.out.print("NAN (8 zenbaki + 1 letra): ");
+            nan = sc.nextLine();
+            if (!Filtroak.isDNI(nan)) {
+                System.out.println(
+                        Gehigarriak.Gorria + "NAN okerra. 8 zenbaki eta 1 letra izan behar ditu." + Gehigarriak.RESET);
+                continue;
+            }
+            if (ErroreenKudeaketa.ifExistsNan(nan, path)) {
+                // NAN existitzen da, beste bat sartu
+                continue;
+            }
+            break; // NAN baliozkoa eta ez dago fitxategian
+        } while (true);
+
+        // Izena balidazioa
+        do {
+            System.out.print("Izena: ");
+            izena = sc.nextLine();
+            if (!Filtroak.isIzena(izena)) {
+                System.out.println(
+                        Gehigarriak.Gorria + "Izena okerra. Letra bakarrik sartu behar da." + Gehigarriak.RESET);
+            }
+        } while (!Filtroak.isIzena(izena));
+
+        // Abizena balidazioa
+        do {
+            System.out.print("Abizena: ");
+            abizena = sc.nextLine();
+            if (!Filtroak.isIzena(abizena)) {
+                System.out.println(
+                        Gehigarriak.Gorria + "Abizena okerra. Letra bakarrik sartu behar da." + Gehigarriak.RESET);
+            }
+        } while (!Filtroak.isIzena(abizena));
+
+        // Adina balidazioa
+        do {
+            System.out.print("Adina: ");
+            adinaStr = sc.nextLine();
+            if (!Filtroak.isAdina(adinaStr)) {
+                System.out.println(
+                        Gehigarriak.Gorria + "Adina okerra. Zenbaki bakarrik sartu behar da." + Gehigarriak.RESET);
+                continue;
+            }
+            adina = Integer.parseInt(adinaStr);
+            break;
+        } while (true);
+
+        // Helbidea
+        System.out.print("Helbidea: ");
+        helbidea = sc.nextLine();
+        helbidea = Filtroak.removeSpaces(helbidea);
+
+        // Pertsona objektua sortu
+        Pertsona p = new Pertsona(nan, izena, abizena, adina, helbidea);
+
+        // Fitxategian gehitu
+        try (BufferedWriter bw = new BufferedWriter(new FileWriter(path, true))) {
+            bw.write(p.toCSV());
+            bw.newLine();
+            System.out.println(Gehigarriak.Berdea + "Datuak ondo gehitu dira fitxategian." + Gehigarriak.RESET);
+        } catch (IOException e) {
             System.out.println(
-                    Gehigarriak.Gorria + "NAN okerra. 8 zenbaki eta 1 letra izan behar ditu." + Gehigarriak.RESET);
-            continue;
+                    Gehigarriak.Gorria + "Errorea: Datuak ezin izan dira gehitu fitxategian." + Gehigarriak.RESET);
+            System.err.println(e.getMessage());
         }
-        if (ErroreenKudeaketa.ifExistsNan(nan, path)) {
-            // NAN existitzen da, beste bat sartu
-            continue;
-        }
-        break; // NAN baliozkoa eta ez dago fitxategian
-    } while (true);
-
-    // Izena balidazioa
-    do {
-        System.out.print("Izena: ");
-        izena = sc.nextLine();
-        if (!Filtroak.isIzena(izena)) {
-            System.out.println(
-                    Gehigarriak.Gorria + "Izena okerra. Letra bakarrik sartu behar da." + Gehigarriak.RESET);
-        }
-    } while (!Filtroak.isIzena(izena));
-
-    // Abizena balidazioa
-    do {
-        System.out.print("Abizena: ");
-        abizena = sc.nextLine();
-        if (!Filtroak.isIzena(abizena)) {
-            System.out.println(
-                    Gehigarriak.Gorria + "Abizena okerra. Letra bakarrik sartu behar da." + Gehigarriak.RESET);
-        }
-    } while (!Filtroak.isIzena(abizena));
-
-    // Adina balidazioa
-    do {
-        System.out.print("Adina: ");
-        adinaStr = sc.nextLine();
-        if (!Filtroak.isAdina(adinaStr)) {
-            System.out.println(
-                    Gehigarriak.Gorria + "Adina okerra. Zenbaki bakarrik sartu behar da." + Gehigarriak.RESET);
-            continue;
-        }
-        adina = Integer.parseInt(adinaStr);
-        break;
-    } while (true);
-
-    // Helbidea
-    System.out.print("Helbidea: ");
-    helbidea = sc.nextLine();
-    helbidea = Filtroak.removeSpaces(helbidea);
-
-    // Pertsona objektua sortu
-    Pertsona p = new Pertsona(nan, izena, abizena, adina, helbidea);
-
-    // Fitxategian gehitu
-    try (BufferedWriter bw = new BufferedWriter(new FileWriter(path, true))) {
-        bw.write(p.toCSV());
-        bw.newLine();
-        System.out.println(Gehigarriak.Berdea + "Datuak ondo gehitu dira fitxategian." + Gehigarriak.RESET);
-    } catch (IOException e) {
-        System.out.println(
-                Gehigarriak.Gorria + "Errorea: Datuak ezin izan dira gehitu fitxategian." + Gehigarriak.RESET);
-        System.err.println(e.getMessage());
     }
-}
 
     private static void txtFitxategiaEguneratu() {
         String fileName, path;
@@ -374,86 +374,96 @@ public class TxtKudeatzailea {
         fileName = sc.next();
         Filtroak.removeSpaces(fileName);
         path = "./fitxategiak/txt/" + fileName + ".txt";
+        String erantzuna;
+
+        do {
+            System.out.print(
+                    Gehigarriak.Horia + "Zihur zaide fitxategia ezabatu nahi duzula? (Bai/Ez) " + Gehigarriak.RESET);
+            erantzuna = sc.next().toLowerCase();
+        } while (erantzuna != "bai" || erantzuna != "ez");
 
         java.nio.file.Path p = java.nio.file.Paths.get(path);
-        try {
-            boolean deleted = java.nio.file.Files.deleteIfExists(p);
-            if (deleted) {
-                System.out.println(Gehigarriak.Berdea + "Fitxategia ondo ezabatu da: " + path + Gehigarriak.RESET);
-            } else {
-                System.out.println(Gehigarriak.Gorria + "Fitxategia ez da aurkitu: " + path + Gehigarriak.RESET);
+
+        if (erantzuna.equals("bai")) {
+            try {
+                boolean deleted = java.nio.file.Files.deleteIfExists(p);
+                if (deleted) {
+                    System.out.println(Gehigarriak.Berdea + "Fitxategia ondo ezabatu da: " + path + Gehigarriak.RESET);
+                } else {
+                    System.out.println(Gehigarriak.Gorria + "Fitxategia ez da aurkitu: " + path + Gehigarriak.RESET);
+                }
+            } catch (IOException e) {
+                System.out.println(Gehigarriak.Gorria + "Errorea: Fitxategia ezin izan da ezabatu.");
+                System.err.println(e.getMessage() + Gehigarriak.RESET);
             }
-        } catch (IOException e) {
-            System.out.println(Gehigarriak.Gorria + "Errorea: Fitxategia ezin izan da ezabatu.");
-            System.err.println(e.getMessage() + Gehigarriak.RESET);
         }
     }
 
     private static void txtFitxategiaCSVraBihurtu() {
         String fileName, path, csvPath;
         Scanner sc = new Gehigarriak().in;
-    
+
         System.out.println("TXT fitxategia CSV formatura bihurtu");
         txtFitxategiakBistaratu();
-    
+
         // Fitxategia aukeratu
         System.out.print(Gehigarriak.Horia + "Zein fitxategi bihurtu nahi duzu? Sartu izena: " + Gehigarriak.RESET);
         fileName = sc.next();
         Filtroak.removeSpaces(fileName);
         path = "./fitxategiak/txt/" + fileName + ".txt";
         csvPath = "./fitxategiak/csv/" + fileName + "-convert_Form_TXT.csv";
-    
+
         // Fitxategia irakurri eta CSVra idatzi
         try (BufferedReader br = new BufferedReader(new FileReader(path));
                 BufferedWriter bw = new BufferedWriter(new FileWriter(csvPath))) {
-    
+
             // CSV-ren kabezalak (Banatzailea: ';')
             String cabezalak = "NAN;Izena;Abizena;Adina;Helbidea";
             bw.write(cabezalak);
             bw.newLine();
-    
+
             // TXT-ren banatzailea (adibidez, espazioa. Aldatu behar baduzu)
             final String TXT_BANATZAILEA = ",";
-            
+
             // CSV-ren banatzailea
             final String CSV_BANATZAILEA = ";";
-    
+
             String lerroa;
             while ((lerroa = br.readLine()) != null) {
-                
+
                 String[] datuak = lerroa.split(TXT_BANATZAILEA);
                 String csvLerroa = "";
-    
+
                 // Egiaztatu lerroak zenbat zutabe dituen
                 if (datuak.length == 2) {
                     // 1. Formatu Laburra: NAN eta Adina (2 zutabe)
                     String nan = datuak[0];
                     String adina = datuak[1];
-                    
+
                     // NAN ; Izena ; Abizena ; Adina ; Helbidea
-                    csvLerroa = nan + CSV_BANATZAILEA + 
-                                "" + CSV_BANATZAILEA +       // Izena (Hutsik)
-                                "" + CSV_BANATZAILEA +       // Abizena (Hutsik)
-                                adina + CSV_BANATZAILEA +    // Adina
-                                "";                          // Helbidea (Hutsik)
-    
+                    csvLerroa = nan + CSV_BANATZAILEA +
+                            "" + CSV_BANATZAILEA + // Izena (Hutsik)
+                            "" + CSV_BANATZAILEA + // Abizena (Hutsik)
+                            adina + CSV_BANATZAILEA + // Adina
+                            ""; // Helbidea (Hutsik)
+
                 } else if (datuak.length == 5) {
                     // 2. Formatu Osoa: NAN, Izena, Abizena, Adina, Helbidea (5 zutabe)
                     // Datuak jada CSV kabezalarekin bat datoz, beraz, berriz muntatu besterik ez.
                     csvLerroa = String.join(CSV_BANATZAILEA, datuak);
-                    
+
                 } else {
                     // Beste formatu bat edo lerro akastuna
                     System.err.println("OHARRA: Lerroak ez du espero den formatua (2 edo 5 zutabe): " + lerroa);
                     continue; // Hurrengo lerrora salto egin
-    
+
                 }
-    
+
                 // CSV fitxategian idatzi
                 bw.write(csvLerroa);
                 bw.newLine();
             }
-            
+
             System.out.println(Gehigarriak.Berdea + "Fitxategia ondo bihurtu da: " + csvPath + Gehigarriak.RESET);
         } catch (FileNotFoundException e) {
             System.out.println("Fitxategia ez da aurkitu.");
